+++ Exodus/src/utopia/exodus/rest/resource/user/me/MeNode.scala
@@ -2,7 +2,7 @@ package utopia.exodus.rest.resource.user.me
 
 import utopia.access.http.Method.Get
 import utopia.access.http.Status.Unauthorized
-import utopia.citadel.database.access.single.user.DbUser
+import utopia.exodus.model.enumeration.ExodusScope.PersonalDataRead
 import utopia.exodus.rest.resource.scalable.{ExtendableSessionResource, SessionUseCaseImplementation}
 import utopia.exodus.rest.util.AuthorizedContext
 import utopia.flow.generic.ValueConversions._
@@ -19,19 +19,28 @@ object MeNode extends ExtendableSessionResource
 {
 	override val name = "me"
 	
-	private val defaultGet = SessionUseCaseImplementation.default(Get) { (session, connection, context, _) =>
+	private val defaultGet = SessionUseCaseImplementation.default { (session, connection, context, _) =>
 		implicit val c: Connection = connection
 		implicit val cntx: AuthorizedContext = context
 		// Reads user data and adds linked data
-		DbUser(session.userId).withLinks match {
-			case Some(user) => Result.Success(user.toModelWith(session.modelStyle))
-			case None =>
-				// Log.warning(s"User id ${session.userId} was authorized but couldn't be found from the database")
-				Result.Failure(Unauthorized, "User no longer exists")
+		session.userAccess match {
+			case Some(userAccess) =>
+				// Makes sure the client is authorized to read user data
+				if (session.access.hasScope(PersonalDataRead))
+					userAccess.withLinks match {
+						case Some(user) => Result.Success(user.toModelWith(session.modelStyle))
+						case None =>
+							// Log.warning(s"User id ${session.userId} was authorized but couldn't be found from the database")
+							Result.Failure(Unauthorized, "User no longer exists")
+					}
+				else
+					Result.Failure(Unauthorized, "You're not allowed to read this user's personal data")
+			
+			case None => Result.Failure(Unauthorized, "Your current session doesn't specify who 'me' is")
 		}
 	}
 	
-	override protected val defaultUseCaseImplementations = Vector(defaultGet)
+	override protected val defaultUseCaseImplementations = Map(Get -> defaultGet)
 	override protected val defaultFollowImplementations =
 		Vector(MySettingsNode, MyOrganizationsNode, MyInvitationsNode, MyLanguagesNode, MyPasswordNode)
 			.map { FollowImplementation.withChild(_) }