+++ Exodus/src/utopia/exodus/rest/resource/user/me/MySessionsNode.scala
@@ -1,0 +1,11 @@
+package utopia.exodus.rest.resource.user.me
+
+/**
+  * Used for interacting with (deviceless) user sessions
+  * @author Mikko Hilpinen
+  * @since 19.2.2022, v4.0
+  */
+object MySessionsNode
+{
+	// TODO: Implement
+}